-Bugs are being cast out! Doom is here for those tiny insects of destruction and mayhem! Their time has come short! Their deaths draw near! If they don't have those little affairs of theirs in order, and pretty soon, those affairs will stay out of order for eternity! So, bugs be ware!

+++ Entitties/EskaadEntities/Sales.cs
@@ -20,9 +20,9 @@ namespace MahtaKala.Entities.EskaadEntities
         public string Validation { get; set; }
 		public string BCode { get; set; }
 		public string BName { get; set; }
-		public int IsInstallmentSale { get; set; }
+		public byte IsInstallmentSale { get; set; }
 		public string BAgentName { get; set; }
 		public string BAgentId { get; set; }
-		public int NumOfInstallments { get; set; }
+		public byte NumOfInstallments { get; set; }
 	}
 }
